Update the agenda item list dropdown menus after create, delete, drop
