+++ modules/meeting/app/components/meeting_agenda_items/item_component/menu_component.html.erb
@@ -1,0 +1,10 @@
+<%=
+  component_wrapper do
+    render(Primer::Alpha::ActionMenu.new) do |menu|
+      menu.with_show_button(icon: "kebab-horizontal", 'aria-label': t("label_agenda_item_actions"), scheme: :invisible, test_selector: 'op-meeting-agenda-actions')
+      edit_action_item(menu) if @meeting_agenda_item.editable?
+      move_actions(menu)
+      delete_action_item(menu)
+    end if edit_enabled?
+  end
+%>
