+++ modules/meeting/app/components/meeting_agenda_items/item_component/show_component.html.erb
@@ -65,12 +65,9 @@
     end
 
     grid.with_area(:actions, tag: :div, justify_self: :end) do
-      render(Primer::Alpha::ActionMenu.new) do |menu|
-        menu.with_show_button(icon: "kebab-horizontal", 'aria-label': t("label_agenda_item_actions"), scheme: :invisible, test_selector: 'op-meeting-agenda-actions')
-        edit_action_item(menu) if @meeting_agenda_item.editable?
-        move_actions(menu)
-        delete_action_item(menu)
-      end if edit_enabled?
+      render(
+        MeetingAgendaItems::ItemComponent::MenuComponent.new(meeting_agenda_item: @meeting_agenda_item)
+      )
     end
 
     grid.with_area(:notes, tag: :div) do
