+++ modules/meeting/app/components/meeting_agenda_items/item_component/show_component.rb
@@ -45,49 +45,6 @@ module MeetingAgendaItems
       @meeting.open? && User.current.allowed_in_project?(:manage_agendas, @meeting.project)
     end
 
-    def edit_enabled?
-      @meeting.open? && User.current.allowed_in_project?(:manage_agendas, @meeting.project)
-    end
-
-    def edit_action_item(menu)
-      menu.with_item(label: t("label_edit"),
-                     href: edit_meeting_agenda_item_path(@meeting_agenda_item.meeting, @meeting_agenda_item),
-                     content_arguments: {
-                       data: { 'turbo-stream': true }
-                     }) do |item|
-        item.with_leading_visual_icon(icon: :pencil)
-      end
-    end
-
-    def move_actions(menu)
-      move_action_item(menu, :highest, t("label_agenda_item_move_to_top"), "move-to-top") unless @meeting_agenda_item.first?
-      move_action_item(menu, :higher, t("label_agenda_item_move_up"), "chevron-up") unless @meeting_agenda_item.first?
-      move_action_item(menu, :lower, t("label_agenda_item_move_down"), "chevron-down") unless @meeting_agenda_item.last?
-      move_action_item(menu, :lowest, t("label_agenda_item_move_to_bottom"), "move-to-bottom") unless @meeting_agenda_item.last?
-    end
-
-    def delete_action_item(menu)
-      menu.with_item(label: t("text_destroy"),
-                     scheme: :danger,
-                     href: meeting_agenda_item_path(@meeting_agenda_item.meeting, @meeting_agenda_item),
-                     form_arguments: {
-                       method: :delete, data: { confirm: t("text_are_you_sure"), 'turbo-stream': true }
-                     }) do |item|
-        item.with_leading_visual_icon(icon: :trash)
-      end
-    end
-
-    def move_action_item(menu, move_to, label_text, icon)
-      menu.with_item(label: label_text,
-                     href: move_meeting_agenda_item_path(@meeting_agenda_item.meeting, @meeting_agenda_item,
-                                                         move_to:),
-                     form_arguments: {
-                       method: :put, data: { 'turbo-stream': true }
-                     }) do |item|
-        item.with_leading_visual_icon(icon:)
-      end
-    end
-
     def duration_color_scheme
       if @meeting.end_time < @meeting_agenda_item.end_time
         :danger
